8825_core/integrations/figjam/requirements.txt
# FigJam Sticky Note Integration
pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
